RH7: hv_utils: drop .getcrosststamp() support from PTP driver
4f9bac039a64f6306b613a0d90e6b7e75d7ab0c4

Turns out that our implementation of .getcrosststamp() never actually
worked. Hyper-V is sending time samples every 5 seconds and this is
too much for get_device_system_crosststamp() as it's interpolation
algorithm (which nobody is currently using in kernel, btw) accounts
for a 'slow' device but we're not slow in Hyper-V, our time reference
is too far away.