RH7: Drivers: hv: vmbus: Perform device register in the per-channel work element
fde25d25dbd997067058f7d4c2ff31600157e6f2

This patch is a continuation of the rescind handling cleanup work. We
cannot
block in the global message handling work context especially if we are
blocking
waiting for the host to wake us up. I would like to thank
Dexuan Cui <decui@microsoft.com> for observing this problem.

The current char-next branch is broken and this patch fixes
the bug.